Refactor -> use parameterized GraphQL query
To make a cleaner separation between graphql and python
code, this commit extracts the graphql query text from the
api_wrapper module.

Folders were restructured to facilitate the future addition
of automated unit tests.

Lastly, a Makefile was added to make it easier to build, test,
and run the burndown chart generator.