Improve date handling
This project was processing dates without proper consideration of
their timezones. As such, several bizarre errors errors were
happening, especially around the start and end of a sprint.

For example, issues closed at 6pm Arizona time on the last day of
the sprint were not having their points credited to that day on
the burndown chart because GitHub returned that closed date as
1am UTC the following day.

This project now processes all dates internally using UTC. The
`sprint_start_date` and `sprint_end_date` are assumed to be
at 00:00:00 *local time*, and the final burndown chart is
rendered in local time.

These changes fix #17 and fix #21